// src/app/api/models/o-aplicacion.ts
/* tslint:disable */
/* eslint-disable */
export interface OAplicacion {
  addAt?: string;
  checkNext?: string;
  checkUlt?: string;
  codigo?: null | string;
  datacenterId?: number;
  descripcion?: null | string;
  editAt?: string;
  estado?: boolean;
  id?: null | string;
  locationId?: number;
  multiServer?: boolean;
  plataformaId?: null | string;
  serverCurrent?: number;
  serverMax?: number;
  serverMin?: number;
  serverType?: number;
  url?: null | string;
}